requirements.txt
torch
torchvision
tqdm
numpy
scikit-learn
transformers
matplotlib
seaborn
Pillow
opencv-python